refactor(knowledge-center): move stripPendingUploadNodes to BlockEditorPureUtils
Pure HTML-manipulation helper belongs alongside the other stateless
BlockEditor content utilities rather than inside the component file.
